added /box endpoints , and new validator middlewares

+++ test/util/test-util.js
@@ -1,18 +1,31 @@
 import mongoose from 'mongoose'
 import config from '../../config.js'
-import fs  from 'fs'
-
-
+import fs from 'fs'
+import Libraries from '../../src/lib/index.js'
+import UseCases from '../../src/use-cases/index.js'
 import SERVER from '../../server.js'
 
+
+const libraries =  new Libraries()
+const useCases = new UseCases({ libraries })
 let APP
 
+
+export const createTestUser = async (inObj = { username : 'test' , password: '1234'})=>{
+  try {
+    const result = await useCases.users.createUser(inObj)
+    return result
+  } catch (error) {
+    console.log(` Error in test/util.js/createTestUser()`, error)
+    throw error
+  }
+}
 
 
 // Remove all collections from the DB.
 export const startApp = async () => {
   if (process.env.ENVIROMENT !== 'test') { throw new Error('Trying to start app without the `test` enviroment') }
-  if(APP) return APP
+  if (APP) return APP
   APP = new SERVER()
   APP.config.storePath = './helia-data-test' // helia data path
   await APP.start()
@@ -38,8 +51,14 @@ export const startDb = async () => {
 // Remove all collections from the DB.
 export const cleanNode = async (nodePath = './helia-data-test') => {
   if (process.env.ENVIROMENT !== 'test') { throw new Error('Trying to remove node without the `test` enviroment') }
- fs.rmdirSync(nodePath ,  { recursive: true})
+  if (!fs.existsSync(nodePath)) {
+    console.log(`test-util/cleanNode info : ${nodePath} not found!`)
+    return
+  }
+  fs.rmdirSync(nodePath, { recursive: true })
 }
+
+
